move disable_notification_dismiss_timer to test_notification_manager
