+++ napari/_qt/_tests/test_qt_notifications.py
@@ -1,7 +1,6 @@
 import threading
 import warnings
 from concurrent.futures import Future
-from contextlib import suppress
 from dataclasses import dataclass
 from unittest.mock import MagicMock
 
@@ -41,26 +40,6 @@ def _threading_raise():
 
 def _raise():
     raise ValueError("error!")
-
-
-@pytest.fixture(autouse=True)
-def disable_notification_dismiss_timer(monkeypatch):
-    """
-    This fixture disables starting timer for closing notification
-    by setting the value of `NapariQtNotification.DISMISS_AFTER` to 0.
-
-    As Qt timer is realised by thread and keep reference to the object,
-    without increase of reference counter object could be garbage collected and
-    cause segmentation fault error when Qt (C++) code try to access it without
-    checking if Python object exists.
-    """
-
-    with suppress(ImportError):
-        from napari._qt.dialogs.qt_notification import NapariQtNotification
-
-        monkeypatch.setattr(NapariQtNotification, "DISMISS_AFTER", 0)
-        monkeypatch.setattr(NapariQtNotification, "FADE_IN_RATE", 0)
-        monkeypatch.setattr(NapariQtNotification, "FADE_OUT_RATE", 0)
 
 
 @pytest.fixture
